+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.1](https://github.com/3xMike/config-manager/releases/tag/0.3.1) - 2025-03-12
 ### Fixed
 - Bug "Can't deserialize ... EoF" on setting empty string via env/cli/config.
+### Updated:
+- `deserialize_with` documentation.
 
 ## [0.3.0](https://github.com/3xMike/config-manager/releases/tag/0.3.0) - 2025-03-12
 ### Changed